fix(bindings): degrees(floatset, bool) + valueN(tgeompoint, int)
Two more individual `mode skip` blocks cleared with targeted
signature fixes.

* **set.test:227** — `SELECT degrees(floatset, true)`.
  `SetFunctions::Floatset_degrees` was reading the second argument
  vector with `FlatVector::GetData<int32_t>(...)` even though the
  binding declared the argument as `BOOLEAN`. DuckDB 1.4 rejects the
  mismatch at bind time ("Expected INT32, found BOOL"). Read the
  column as `bool` to match the registered type.

* **tgeompoint.test:482** — `SELECT ST_AsText(valueN(tgeompoint, 1))`.
  The `valueN` overload on TGEOMPOINT declared its second argument as
  `LogicalType::INTEGER` while every sibling `valueN` overload on
  other temporal types declared `LogicalType::BIGINT`, and the shared
  C implementation `TemporalFunctions::Temporal_value_n` uses
  `BinaryExecutor<string_t, int64_t, string_t>`. With the tgeompoint
  overload the Vector was INT32 but the template wanted INT64, hence
  "Expected INT64, found INT32". Align with BIGINT like the others.

Both test blocks are un-wrapped from `mode skip` / `mode unskip` now
that they pass. Remaining bump-PR skips: the three wholesale-skipped
files (spanset.test, geoset.test, tbox.test) — addressed in subsequent
commits on this PR.

+++ src/geo/tgeompoint.cpp
@@ -398,7 +398,12 @@ void TgeompointType::RegisterScalarFunctions(ExtensionLoader &loader) {
     loader.RegisterFunction(
         ScalarFunction(
             "valueN",
-            {TGEOMPOINT(), LogicalType::INTEGER},
+            // BIGINT to match the sibling `valueN(<other temporal>, BIGINT)`
+            // overload in temporal.cpp and the C function's int64_t template
+            // arg (BinaryExecutor<string_t,int64_t,string_t>). Registering
+            // INTEGER here made DuckDB 1.4 reject the bind with
+            // "Expected INT64, found INT32" (tgeompoint.test:482).
+            {TGEOMPOINT(), LogicalType::BIGINT},
             GeoTypes::GEOMETRY(),
             TemporalFunctions::Temporal_value_n
         )
